+++ storage/s3/src/server/internal/Helpers.ts
@@ -2,10 +2,9 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import {
-  defaultExpiresInSeconds,
-  ExpiryOptions,
-} from "@itwin/object-storage-core/lib/common/internal";
+import { defaultExpiresInSeconds } from "@itwin/object-storage-core/lib/common/internal";
+
+import { ExpiryOptions } from "@itwin/object-storage-core";
 
 import { S3ClientWrapper } from "../wrappers";
