+++ tests/backend-storage-unit/src/shared/Mocks.ts
@@ -2,9 +2,9 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { ExpiryOptions } from "@itwin/object-storage-core/lib/common/internal";
 
 import {
+  ExpiryOptions,
   ObjectDirectory,
   ObjectReference,
   PresignedUrlProvider,
