+++ tests/backend-storage/src/test/ServerStorage.test.ts
@@ -8,15 +8,13 @@ import * as path from "path";
 import { expect, use } from "chai";
 import * as chaiAsPromised from "chai-as-promised";
 
-import {
-  defaultExpiresInSeconds,
-  ExpiryOptions,
-} from "@itwin/object-storage-core/lib/common/internal";
+import { defaultExpiresInSeconds } from "@itwin/object-storage-core/lib/common/internal";
 import { getRandomString } from "@itwin/object-storage-core/lib/server/internal";
 
 import {
   BaseDirectory,
   ContentHeaders,
+  ExpiryOptions,
   Metadata,
   ObjectReference,
   ServerStorage,
